Move all manual test for recipe manager into derived class
Invoke test through test_recipe_manager.py
