+++ recipemanager.py
@@ -1,41 +1,16 @@
 from ingredient import Ingredient
 from recipe import Recipe
+from typing import List
 
 class RecipeManager:
     def __init__(self):
-        self.recipes = []
-
-    def display_recipes(self):
-        for recipe in self.recipes:
-            print(recipe._title) #title instead the name 
-            return 
-        if not self.recipes:
-            print("No recipes to display.")
-            return
-
+        self.recipes:List[Recipe] = []
 
     def add_recipe(self, recipe_to_add: Recipe):
-        self.recipes.append(recipe_to_add)
-    def add_recipe_from_input(self):
-        title = input("Enter the title of the dish: ")
-        description = input("Enter the description: ")
-        no_servings = int(input("Enter the number of servings: "))
-        calories_per_portion = int(input("Enter the calories per portion: "))
-        ingredients = input("Enter the ingredients: ").split(",")
-        instructions = input("Enter the instructions: ").split(",")
-
-        new_recipe = Recipe(title)
-        new_recipe._description = description  # Set description directly
-        new_recipe.set_no_servings(no_servings)
-        new_recipe.set_calories_per_portion(calories_per_portion)
-        for ingredient in ingredients:
-            new_recipe.set_ingredients(ingredient.strip())
-        for instruction in instructions:
-            new_recipe.set_instructions(instruction.strip())
-        
-        self.add_recipe(new_recipe)
-        print("Recipe added.")
-        
+        if len(self.recipes) > 0:
+            self.recipes.append(recipe_to_add)
+        else:
+            self.recipes = [recipe_to_add]
 
     def update_recipe(self, recipe_to_update: Recipe, updated_recipe: Recipe):
         recipe_to_update.set_description(updated_recipe.get_description())
@@ -43,67 +18,6 @@ class RecipeManager:
         recipe_to_update.set_calories_per_portion(updated_recipe.get_calories_per_portion())
         recipe_to_update.set_ingredients(updated_recipe.get_ingredients())
         recipe_to_update.set_instructions(updated_recipe.get_instructions())
-    def update_recipe_from_input(self):
-        title_to_update = input("Enter the name of the recipe to update: ")
-        found = False
-        for recipe in self.recipes:
-            if recipe._title == title_to_update:
-                description = input("Enter the updated description: ")
-                no_servings = int(input("Enter the updated number of servings: "))
-                calories_per_portion = int(input("Enter the updated calories per portion: "))
-                ingredients = input("Enter the updated ingredients: ").split(",")
-                instructions = input("Enter the updated instructions: ").split(",")
-                updated_recipe = Recipe(title_to_update)
-                updated_recipe.set_description(description)
-                updated_recipe.set_no_servings(no_servings)
-                updated_recipe.set_calories_per_portion(calories_per_portion)
-                for ingredient in ingredients:
-                    updated_recipe.set_ingredients(ingredient.strip())
-                for instruction in instructions:
-                    updated_recipe.set_instructions(instruction.strip())
-
-                self.update_recipe(recipe, updated_recipe)
-                print("Recipe updated successfully.")
-                found = True
-                break
-
-        if not found:
-            print("Recipe not found.")
-        
-
-
     
     def delete_recipe(self, recipe_to_delete: Recipe):
         self.recipes.remove(recipe_to_delete)
-    def delete_recipe_from_input(self):
-        title_to_delete = input("Enter the name of the recipe to delete: ")
-        found = False
-
-        for recipe in self.recipes:
-            if recipe._title == title_to_delete:
-                self.delete_recipe(recipe)
-                print("Recipe deleted successfully.")
-                found = True
-                break
-
-        if not found:
-            print("Recipe not found. Unable to delete.")
-
-
-    def menu(self):
-        print("1. Do you want to display recipes?")
-        print("2. Do you want to add recipes?")
-        print("3. Do you want to update recipe?")
-        print("4. Do you want to delete recipe?")
-        choice = input("Enter the number of your choice: ")
-
-        if choice == "1":
-            self.display_recipes()
-        elif choice == "2":
-            self.add_recipe_from_input()
-        elif choice == "3":
-            self.update_recipe_from_input()
-        elif choice == "4":
-            self.delete_recipe_from_input()
-        else:
-            print("Invalid choice.")
