+++ test/test_recipe_manager.py
@@ -1,0 +1,6 @@
+from TestRecipeManager import TestRecipeManager
+
+testrecipemanager = TestRecipeManager()
+
+while True:
+   testrecipemanager.menu()
